Update install-guide.md
There a few issues which need to be addressed. 
The first being  that the TOC is not up to date and several headings are not shown in the Table of Contents. I did not make that update.

Another is the section titled First Start (line # 226 through #330):
I made corrections to some of the text to help make the text easier to understand. However, after reading through the entire section it seems to be out of order. I would suggest further that the title be changed to something like "Launching the Application for the First Time" or "Initialize the Application".....
If you see as I do that this section and the guide in general is out of order. I would further suggest that you start with a Setup section.