Add Secret Key to admin/configuration.md
Resolves: #115

Not setting the HTTP Secret Key is a common issue encountered by users
installing the application - particularly via a package managager (e.g.
debian apt).

By putting this as the first section of the configuration guide, we
should be able to reduce the number of support requests and issues
raised from users encountering this issue.